+++ src/Core/Dfe.Complete.Application/Common/Models/Result.cs
@@ -4,6 +4,7 @@
     {
         None,
         Unauthorized,
+        NotFound,
         Unknown
     }