+++ src/Core/Dfe.Complete.Application/Notes/Commands/UpdateNoteCommand.cs
@@ -1,10 +1,9 @@
 using Dfe.Complete.Application.Common.Models;
 using Dfe.Complete.Application.Notes.Interfaces;
 using Dfe.Complete.Application.Projects.Models;
+using Dfe.Complete.Application.Notes.Queries.QueryFilters;
 using Dfe.Complete.Domain.ValueObjects;
-using Dfe.Complete.Utils;
 using MediatR;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
 namespace Dfe.Complete.Application.Notes.Commands;
@@ -12,8 +11,8 @@ namespace Dfe.Complete.Application.Notes.Commands;
 public record UpdateNoteCommand(NoteId NoteId, string Body) : IRequest<Result<NoteDto>>;
 
 public class UpdateNoteCommandHandler(
-    INoteWriteRepository _repo,
-    IHttpContextAccessor _httpContextAccessor,
+    INoteWriteRepository _noteWriteRepo,
+    INoteReadRepository _noteReadRepo,
     ILogger<UpdateNoteCommandHandler> logger
 ) : IRequestHandler<UpdateNoteCommand, Result<NoteDto>>
 {
@@ -21,19 +20,14 @@ public class UpdateNoteCommandHandler(
     {
         try
         {
-            var userIdClaim = _httpContextAccessor.HttpContext?.User?.FindFirst("sub")?.Value
-                ?? _httpContextAccessor.HttpContext?.User?.FindFirst("uid")?.Value;
+            var note = new NoteIdQuery(request.NoteId)
+                .Apply(_noteReadRepo.Notes())
+                .FirstOrDefault();
 
-            if (!Guid.TryParse(userIdClaim, out var parsedUserId))
-                throw new UnauthorizedAccessException($"Could not update note {request.NoteId.Value}");
-
-            var note = await _repo.GetNoteByIdAsync(request.NoteId, cancellationToken) ?? throw new NotFoundException($"Note with ID {request.NoteId.Value} not found");
-
-            if (note.UserId != new UserId(parsedUserId))
-                throw new UnauthorizedAccessException($"Could not update note {request.NoteId.Value}");
+            if (note is null) return Result<NoteDto>.Failure($"Note with ID {request.NoteId.Value} not found", ErrorType.NotFound);
 
             note.Body = request.Body;
-            await _repo.UpdateNoteAsync(note, cancellationToken);
+            await _noteWriteRepo.UpdateNoteAsync(note, cancellationToken);
 
             var noteDto = new NoteDto(
                 note.Id,
@@ -46,11 +40,6 @@ public class UpdateNoteCommandHandler(
             );
 
             return Result<NoteDto>.Success(noteDto);
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            logger.LogWarning(ex, "Exception for {Name} Request - {@Request}", nameof(RemoveNoteCommandHandler), request);
-            return Result<NoteDto>.Failure(ex.Message, ErrorType.Unauthorized);
         }
         catch (Exception ex)
         {